requirements.txt
Flask==3.0.0
werkzeug==3.0.0
torch==2.1.0
openai-whisper==20231117
transformers==4.35.0
sentencepiece==0.1.99
protobuf==3.20.3
gTTS==2.4.0
numpy==1.24.3
